pyproject.toml
[tool.poetry]
name = "connect-notion-and-clockify"
version = "0.1.0"
description = ""
authors = ["koroku1023 <h1023skoroku@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.10"
pandas = "^2.1.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
